requirements.txt
Flask~=2.2.2

# Flask Extensions
Flask-WTF~=1.0.1
Flask-Login~=0.6.2
Flask-Session~=0.4.0

# Utilities
requests~=2.28.1
python-box~=6.0.2

# Dev Dependencies
djlint~=1.14.2
yapf~=0.32.0
toml~=0.10.2